Fix props names

+++ src/app/components/buttons/UserPageButtons.jsx
@@ -8,13 +8,13 @@ export const LogoutButton = (props) => {
   function logoutClicked() {
     logout().then(
       () => {
-        props.err("");
+        props.setConnErr("");
         localStorage.clear();
-        props.msg("You have been successfully logged out!");
+        props.success("You have been successfully logged out!");
       },
       (error) => {
-        props.err(error.message);
-        props.msg("");
+        props.setConnErr(error.message);
+        props.success("");
       }
     );
   }
@@ -43,17 +43,24 @@ export const DeleteAccountButton = (props) => {
   const [question, setQuestion] = useState(false);
 
   function deleteAccount() {
-    deleteUser(currentUser).then(
-      () => {
-        props.err("");
-        localStorage.clear();
-        props.msg("Account has been successfully deleted!");
-      },
-      (error) => {
-        props.err(error.message);
-        props.msg("");
-      }
-    );
+    props.setConnErr("");
+    deleteUser(currentUser).catch((error) => {
+      props.setConnErr(error.message);
+      props.success("");
+    });
+
+    if (props.connErr !== "") return;
+
+    const options = {
+      method: "POST",
+      headers: { "Content-Type": "application/json;charset=utf-8" },
+      body: JSON.stringify({ id: currentUser.uid }),
+    };
+
+    fetch("http://127.0.0.1:5000/api/delete-user", options).then(() => {
+      localStorage.clear();
+      props.success("Account has been successfully deleted!");
+    });
   }
 
   return (
